requirements.txt
CouchDB==1.2
gcloud==0.17.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
jws==0.1.3
mysqlclient==2.0.1
oauth2client==3.0.0
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pymongo==3.11.0
Pyrebase==3.0.27
python-jwt==2.0.1
redis==3.5.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
